add multi iteration mode

+++ main.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	"io/ioutil"
 	"log"
+	"math"
 	"os"
 	"sort"
 	"strconv"
@@ -50,34 +51,88 @@ import (
 
 func main() {
 	var source []byte
-	var cfg combat.Profile
+
 	var err error
 
-	debugPtr := flag.String("d", "error", "output level: debug, info, warn")
-	secondsPtr := flag.Int("s", 600, "how many seconds to run the sim for")
-	pPtr := flag.String("p", "config.yaml", "which profile to use")
+	debug := flag.String("d", "error", "output level: debug, info, warn")
+	seconds := flag.Int("s", 120, "how many seconds to run the sim for")
+	cfgFile := flag.String("p", "config.yaml", "which profile to use")
 	f := flag.String("o", "", "detailed log file")
 	hp := flag.Float64("hp", 0, "hp mode: how much hp to deal damage to")
 	showCaller := flag.Bool("c", false, "show caller in debug low")
-	flag.Parse()
+	avgMode := flag.Bool("a", false, "run sim multiple times and calculate avg damage (smooth out randomness). default false. note that there is no debug log in this mode")
+	w := flag.Int("w", 24, "number of workers to run when running multiple iterations; default 24")
+	i := flag.Int("i", 200, "number of iterations to run if we're running multiple")
 
-	defer profile.Start(profile.ProfilePath("./")).Stop()
+	flag.Parse()
 
 	// p := "./xl-base.yaml" //xl.yaml expecting 4659 dps
 
-	source, err = ioutil.ReadFile(*pPtr)
-	if err != nil {
-		log.Fatal(err)
-	}
-	err = yaml.Unmarshal(source, &cfg)
+	source, err = ioutil.ReadFile(*cfgFile)
 	if err != nil {
 		log.Fatal(err)
 	}
 
-	cfg.LogConfig.LogLevel = *debugPtr
-	cfg.LogConfig.LogFile = *f
-	cfg.LogConfig.LogShowCaller = *showCaller
-	os.Remove(*f)
+	if *avgMode {
+		runIter(*i, *w, source, *hp, *seconds)
+	} else {
+		defer profile.Start(profile.ProfilePath("./")).Stop()
+		var cfg combat.Profile
+		err = yaml.Unmarshal(source, &cfg)
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		cfg.LogConfig.LogLevel = *debug
+		cfg.LogConfig.LogFile = *f
+		cfg.LogConfig.LogShowCaller = *showCaller
+		os.Remove(*f)
+		runSingle(cfg, *hp, *seconds)
+	}
+
+}
+
+func graphToCSV(in []float64) {
+	os.Remove("graph.csv")
+	file, err := os.Create("result.csv")
+	if err != nil {
+		log.Panic(err)
+	}
+	defer file.Close()
+
+	writer := csv.NewWriter(file)
+	defer writer.Flush()
+	t := 10
+	l := len(in) / (60 * t)
+
+	out := make([][]string, 0, l)
+
+	stop := false
+	next := 0
+	var prev float64
+	var i int64
+	for !stop {
+		val := in[next] - prev
+		prev = in[next]
+		out = append(out, []string{strconv.FormatInt(i, 10), strconv.FormatFloat(val/float64(t), 'f', 2, 64)})
+		i++
+		if next == len(in)-1 {
+			stop = true
+		} else {
+			next += 600
+			if next >= len(in) {
+				next = len(in) - 1
+			}
+		}
+	}
+
+	for _, v := range out {
+		writer.Write(v)
+	}
+
+}
+
+func runSingle(cfg combat.Profile, hp float64, dur int) {
 
 	s, err := combat.New(cfg)
 	if err != nil {
@@ -87,13 +142,12 @@ func main() {
 	var stats combat.SimStats
 	var dmg float64
 
-	dur := *secondsPtr
 	start := time.Now()
-	if *hp > 0 {
-		dmg, stats = s.RunHPMode(*hp)
+	if hp > 0 {
+		dmg, stats = s.RunHPMode(hp)
 		dur = stats.SimDuration / 60
 	} else {
-		dmg, stats = s.Run(*secondsPtr)
+		dmg, stats = s.Run(dur)
 	}
 	elapsed := time.Since(start)
 	fmt.Println("------------------------------------------")
@@ -154,50 +208,107 @@ func main() {
 		fmt.Printf("%v: %v\n", k, v)
 	}
 	fmt.Println("------------------------------------------")
-	fmt.Printf("Running profile %v, total damage dealt: %.2f over %v seconds. DPS = %.2f. Sim took %s\n", *pPtr, dmg, dur, dmg/float64(dur), elapsed)
+	fmt.Printf("Running profile %v, total damage dealt: %.2f over %v seconds. DPS = %.2f. Sim took %s\n", cfg.Label, dmg, dur, dmg/float64(dur), elapsed)
 
-	if *hp == 0 {
+	if hp == 0 {
 		graphToCSV(stats.DamageHist)
 	}
-
 }
 
-func graphToCSV(in []float64) {
-	os.Remove("graph.csv")
-	file, err := os.Create("result.csv")
-	if err != nil {
-		log.Panic(err)
+func runIter(n, w int, src []byte, hp float64, dur int) {
+	var progress, sum, ss, min, max float64
+	var data []float64
+	min = math.MaxFloat64
+	max = -1
+
+	count := n
+
+	resp := make(chan float64, n)
+	req := make(chan bool)
+	done := make(chan bool)
+
+	for i := 0; i < w; i++ {
+		go worker(src, hp, dur, resp, req, done)
 	}
-	defer file.Close()
 
-	writer := csv.NewWriter(file)
-	defer writer.Flush()
-	t := 10
-	l := len(in) / (60 * t)
+	go func() {
+		var wip int
+		for wip < n {
+			req <- true
+			wip++
+		}
+	}()
 
-	out := make([][]string, 0, l)
+	fmt.Print("\tProgress: 0")
 
-	stop := false
-	next := 0
-	var prev float64
-	var i int64
-	for !stop {
-		val := in[next] - prev
-		prev = in[next]
-		out = append(out, []string{strconv.FormatInt(i, 10), strconv.FormatFloat(val/float64(t), 'f', 2, 64)})
-		i++
-		if next == len(in)-1 {
-			stop = true
-		} else {
-			next += 600
-			if next >= len(in) {
-				next = len(in) - 1
-			}
+	for count > 0 {
+		val := <-resp
+		count--
+		data = append(data, val)
+		sum += val
+		if val < min {
+			min = val
+		}
+		if val > max {
+			max = val
+		}
+		if (1 - float64(count)/float64(n)) > (progress + 0.1) {
+			progress = (1 - float64(count)/float64(n))
+			fmt.Printf(".%.0f", 100*progress)
 		}
 	}
+	fmt.Print("...done\n")
 
-	for _, v := range out {
-		writer.Write(v)
+	close(done)
+
+	mean := sum / float64(n)
+
+	for _, v := range data {
+		ss += (v - mean) * (v - mean)
 	}
 
+	sd := math.Sqrt(ss / float64(n))
+
+	fmt.Printf(
+		"Simulation done; %v iterations resulted in average dps of %0.0f over %v seconds (min: %0.02f, max: %0.02f, stddev: %0.02f)\n",
+		n,
+		mean,
+		dur,
+		min,
+		max,
+		sd,
+	)
+}
+
+func worker(src []byte, hp float64, dur int, resp chan float64, req chan bool, done chan bool) {
+
+	for {
+		select {
+		case <-req:
+			var cfg combat.Profile
+			err := yaml.Unmarshal(src, &cfg)
+			if err != nil {
+				log.Fatal(err)
+			}
+			cfg.LogConfig.LogLevel = "error"
+			cfg.LogConfig.LogFile = ""
+			cfg.LogConfig.LogShowCaller = false
+
+			s, err := combat.New(cfg)
+			if err != nil {
+				log.Fatal(err)
+			}
+
+			if hp > 0 {
+				dmg, _ := s.RunHPMode(hp)
+				resp <- dmg / float64(dur)
+			} else {
+				dmg, _ := s.Run(dur)
+				resp <- dmg / float64(dur)
+			}
+
+		case <-done:
+			return
+		}
+	}
 }
